refactor: only run watcher in development

+++ packages/@averjs/renderer/lib/renderer.js
@@ -41,32 +41,34 @@ export default class Renderer {
 
     for (const templateFile of templates) this.writeTemplateFile(templateFile);
 
-    const watcher = chokidar.watch(
-      // generate a new set of unique paths
-      [ ...new Set(this.config.templates.map(temp => path.resolve(temp.pluginPath, './entries'))) ]
-    );
-          
-    watcher.on('ready', () => {
-      watcher.on('all', (event, id) => {
-        if (event !== 'addDir' && event !== 'unlinkDir') {
-          let template = this.config.templates.find(temp => temp.src === id);
-          if (!template) {
-            // Try to find any entry file from same plugin to get the plugin path
-            const { dirname } = this.config.templates
-              .find(temp => !path.relative(path.resolve(temp.pluginPath, './entries'), id).startsWith('..'));
-            const dst = path.relative(dirname, id).replace('entries', dirname);
-  
-            template = { src: id, dst };
-  
-            // Push the newly created entry template into the config templates array so we dont have to construct the path again later
-            this.config.templates.push(template);
+    if (!this.isProd) {
+      const watcher = chokidar.watch(
+        // generate a new set of unique paths
+        [ ...new Set(this.config.templates.map(temp => path.resolve(temp.pluginPath, './entries'))) ]
+      );
+            
+      watcher.on('ready', () => {
+        watcher.on('all', (event, id) => {
+          if (event !== 'addDir' && event !== 'unlinkDir') {
+            let template = this.config.templates.find(temp => temp.src === id);
+            if (!template) {
+              // Try to find any entry file from same plugin to get the plugin path
+              const { dirname } = this.config.templates
+                .find(temp => !path.relative(path.resolve(temp.pluginPath, './entries'), id).startsWith('..'));
+              const dst = path.relative(dirname, id).replace('entries', dirname);
+    
+              template = { src: id, dst };
+    
+              // Push the newly created entry template into the config templates array so we dont have to construct the path again later
+              this.config.templates.push(template);
+            }
+    
+            if (event === 'unlink') fs.unlinkSync(path.resolve(this.cacheDir, template.dst));
+            else this.writeTemplateFile(template);
           }
-  
-          if (event === 'unlink') fs.unlinkSync(path.resolve(this.cacheDir, template.dst));
-          else this.writeTemplateFile(template);
-        }
+        });
       });
-    });
+    }
   }
 
   writeTemplateFile(templateFile) {
